#3665 Refactoring: Move skip related stuff in skip module. Removed useless `into()`
