+++ src/utils.rs
@@ -623,54 +623,6 @@ pub(crate) fn unicode_str_width(s: &str) -> usize {
     s.width()
 }
 
-#[derive(Default, Clone)]
-pub(crate) struct SkipContext {
-    macros: Vec<String>,
-    attributes: Vec<String>,
-}
-
-impl SkipContext {
-    pub(crate) fn update_by_attrs(&mut self, attrs: &[ast::Attribute]) {
-        self.macros.append(&mut get_skip_names("macros", attrs));
-        self.attributes
-            .append(&mut get_skip_names("attributes", attrs));
-    }
-
-    pub(crate) fn update(&mut self, mut other: SkipContext) {
-        self.macros.append(&mut other.macros);
-        self.attributes.append(&mut other.attributes);
-    }
-
-    pub(crate) fn macro_skip(&self, name: &str) -> bool {
-        self.macros.iter().any(|n| n == name)
-    }
-
-    pub(crate) fn attributes_skip(&self, name: &str) -> bool {
-        self.attributes.iter().any(|n| n == name)
-    }
-}
-
-fn get_skip_names(kind: &str, attrs: &[ast::Attribute]) -> Vec<String> {
-    let mut skip_names = vec![];
-    let path = format!("rustfmt::skip::{}", kind);
-    for attr in attrs {
-        // syntax::ast::Path is implemented partialEq
-        // but it is designed for segments.len() == 1
-        if format!("{}", attr.path) != path {
-            continue;
-        }
-
-        if let Some(list) = attr.meta_item_list() {
-            for nested_meta_item in list {
-                if let Some(name) = nested_meta_item.ident() {
-                    skip_names.push(name.to_string());
-                }
-            }
-        }
-    }
-    skip_names
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
